Modify evidence formats
Adjusting the evidence formats to be less generic and more tailored to
our use-case.
The WebAuthn-derived TPM formats are now Parsec-centric, selectively
carrying what will be needed for Veraison to properly appraise (both key
and platform attestation). Platforms are identified via their attesting
keys, by using a key identifier (and not an X.509 cert chain).

The user data included in the attestation token consists of only the
nonce coming from the relying party.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>